Rename stubs and update tests
